Removing unused code and simplify init logic
This PR removes multiple unused parts of the code:
- BackendType was tracked everywhere for no reasons.
- Multiple functions were public while not used outside of their module.
- Some functions, for azure, were no longer used.
- Some code duplication when parsing the backend.
- Some struct like 'GenericConnector' was unnecessary.
- Migrating session.go code within the code logic to make the code easier to read and follow.
