+++ backend/hashicorp/vault_test.go
@@ -171,7 +171,7 @@ func TestNewVaultConfigFromBackendConfig_AWSAuth(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			auth, err := NewVaultConfigFromBackendConfig(tt.sessionConfig)
+			auth, err := newVaultConfigFromBackendConfig(tt.sessionConfig)
 
 			if tt.expectError {
 				assert.Error(t, err)
